requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
loguru==0.7.2
google-generativeai==0.3.2
tiktoken==0.5.1 